+++ source/Api/Program.cs
@@ -17,6 +17,7 @@ using System.IdentityModel.Tokens.Jwt;
 using System.Linq;
 using System.Threading.Tasks;
 using BuildingBlocks.Application.Configuration.Logging;
+using BuildingBlocks.Application.Extensions.DependencyInjection;
 using Energinet.DataHub.Core.App.FunctionApp.Extensions.DependencyInjection;
 using Energinet.DataHub.EDI.Api.Authentication;
 using Energinet.DataHub.EDI.Api.Authentication.Certificate;
@@ -115,11 +116,12 @@ namespace Energinet.DataHub.EDI.Api
 
                             return correlationContext;
                         });
-                    services.AddLiveHealthCheck();
-                    services.AddExternalDomainServiceBusQueuesHealthCheck(
-                        runtime.SERVICE_BUS_CONNECTION_STRING_FOR_DOMAIN_RELAY_MANAGE!,
-                        runtime.EDI_INBOX_MESSAGE_QUEUE_NAME!,
-                        runtime.WHOLESALE_INBOX_MESSAGE_QUEUE_NAME!);
+                    services.AddLiveHealthCheck()
+                        .AddExternalDomainServiceBusQueuesHealthCheck(
+                            runtime.SERVICE_BUS_CONNECTION_STRING_FOR_DOMAIN_RELAY_MANAGE!,
+                            runtime.EDI_INBOX_MESSAGE_QUEUE_NAME!,
+                            runtime.WHOLESALE_INBOX_MESSAGE_QUEUE_NAME!)
+                        .AddSqlServerHealthCheck(configuration);
                     services.AddBlobStorageHealthCheck("edi-web-jobs-storage", runtime.AzureWebJobsStorage!);
                     services.AddBlobStorageHealthCheck("edi-documents-storage", runtime.AZURE_STORAGE_ACCOUNT_URL!);
