+++ source/B2CWebApi/Extensions/DependencyInjection/DateTimeExtensions.cs
@@ -25,7 +25,10 @@ public static class DateTimeExtensions
     {
         services
             .AddOptions<DateTimeOptions>()
-            .Bind(configuration);
+            .Bind(configuration)
+            .Validate(
+                o => !string.IsNullOrEmpty(o.TIME_ZONE),
+                "TIME_ZONE must be set");
 
         var options = configuration.Get<DateTimeOptions>()!;
         services.AddSingleton<DateTimeZone>(_ =>
