+++ source/BuildingBlocks.Application/BuildingBlocks.Application.csproj
@@ -10,6 +10,7 @@
       <PackageReference Include="AspNetCore.HealthChecks.AzureServiceBus" Version="8.0.0" />
       <PackageReference Include="AspNetCore.HealthChecks.SqlServer" Version="8.0.0" />
       <PackageReference Include="Azure.Messaging.ServiceBus" Version="7.17.2" />
+      <PackageReference Include="Energinet.DataHub.Core.App.FunctionApp" Version="8.1.0" />
       <PackageReference Include="Microsoft.ApplicationInsights" Version="2.22.0" />
       <PackageReference Include="Microsoft.Extensions.DependencyInjection.Abstractions" Version="8.0.0" />
       <PackageReference Include="Microsoft.Extensions.Options.ConfigurationExtensions" Version="8.0.0" />
